+++ ash/.ashrc.d/21_boot2docker.sh
@@ -176,6 +176,14 @@ b2d_dk_irestore(){
 
 b2d_dk_iremove(){
 
+    # Remove backups only
+    if [[ "${1}" == "backups" ]]; then
+        echo "[INFO] [$(date +"%T")] Removing images backups !"
+        rm -rvf "${BOOT2DOCKER_DK_IMAGES_SAVE_DIR}"
+         echo "[INFO] [$(date +"%T")] Images backup remove completed !"
+        return 0
+    fi
+
     # Done !
     echo "[INFO] [$(date +"%T")] Will remove all your images !"
 
@@ -186,7 +194,7 @@ b2d_dk_iremove(){
     done
 
     # Done !
-    echo "[INFO] [$(date +"%T")] Remove completed !"
+    echo "[INFO] [$(date +"%T")] Images remove completed !"
 }
 
 
@@ -199,7 +207,7 @@ b2d_dk(){
             ;;
         irestore) b2d_dk_irestore
             ;;
-        iremove) b2d_dk_iremove
+        iremove) b2d_dk_iremove "$2"
             ;;
         proxyd) b2d_dk_proxyd "$2"
             ;;
@@ -214,10 +222,10 @@ b2d_dk_custom_usage(){
     echo "b2d dk = commands to manage your docker daemon in your boot2docker environment."
     echo
     echo "b2d dk commands:"
-    echo "    ibackup     Backup all your b2d docker images to a folder in your project"
-    echo "    irestore    Restore all your b2d docker images from a folder in your project"
-    echo "    iremove     Remove all your b2d docker images (dk rmi on all dk images)"
-    echo "    proxyd      Start or stop transparent proxy in a container for your containers"
+    echo "    ibackup             Backup all your b2d docker images to a folder in your project"
+    echo "    irestore            Restore all your b2d docker images from a folder in your project"
+    echo "    iremove [backups]   Remove all your b2d docker images (dk rmi on all dk images) or backups (ibackup)"
+    echo "    proxyd              Start or stop transparent proxy in a container for your containers"
 }